Implement required native functions from mruby's `string.c`
The mruby VM requires several routines from `string.c` that are not used
for implementing Ruby `String` methods. Reimplement these in Rust as
`#[no_mangle] unsafe extern "C" fn`s or in `mruby-sys` by copy-and-pasting
methods from mruby.

This commit adds several `#ifndef ARTICHOKE` blocks to `string.c` to
disable the vast majority of the mruby string source when building for
Artichoke. This approach is similar to the approach taken for `Array`.

Artichoke does not link as of this commit.